+++ plone/app/widgets/tests/test_utils.py
@@ -4,6 +4,7 @@ from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
 from plone.app.widgets.testing import PLONEAPPWIDGETS_INTEGRATION_TESTING
 from plone.app.widgets.utils import get_relateditems_options
+from plone.app.widgets.utils import get_tinymce_options
 
 import unittest
 
@@ -34,15 +35,15 @@ class UtilsTests(unittest.TestCase):
                            # works, even if it was imported before.,,
             orig_HAS_PAE = utils.HAS_PAE
             utils.HAS_PAE = True
-            self.assertEquals(utils.first_weekday(), 0)
+            self.assertEqual(utils.first_weekday(), 0)
             base.first_weekday = lambda: 1
-            self.assertEquals(utils.first_weekday(), 1)
+            self.assertEqual(utils.first_weekday(), 1)
             base.first_weekday = lambda: 5
-            self.assertEquals(utils.first_weekday(), 1)
+            self.assertEqual(utils.first_weekday(), 1)
 
             # test without plone.app.event installed
             utils.HAS_PAE = False
-            self.assertEquals(utils.first_weekday(), 0)
+            self.assertEqual(utils.first_weekday(), 0)
 
         # restore original state
         utils.HAS_PAE = orig_HAS_PAE
@@ -90,32 +91,32 @@ class TestRelatedItemsOptions(unittest.TestCase):
         # context_url contains something, otherwise this test is meaningless
         self.assertTrue(bool(context_url))
 
-        self.assertEquals(
+        self.assertEqual(
             options['rootUrl'],
             root_url
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['rootPath'],
             root_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['vocabularyUrl'],
             root_url + '/@@vocab?name=test_vocab&field=testfield'
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['basePath'],
             context_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['contextPath'],
             context_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['separator'],
             '#!@'
         )
@@ -162,42 +163,42 @@ class TestRelatedItemsOptions(unittest.TestCase):
         # context_url contains something, otherwise this test is meaningless
         self.assertTrue(bool(context_url))
 
-        self.assertEquals(
+        self.assertEqual(
             options['rootUrl'],
             root_url
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['rootPath'],
             root_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['vocabularyUrl'],
             root_url + '/@@vocab?name=test_vocab&field=testfield'
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['basePath'],
             context_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['contextPath'],
             context_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['separator'],
             '#!@'
         )
 
-        self.assertEquals(
+        self.assertEqual(
             len(options['favorites']),
             2
         )
 
-        self.assertEquals(
+        self.assertEqual(
             sorted(options['favorites'][0].keys()),
             ['path', 'title']
         )
@@ -241,32 +242,32 @@ class TestRelatedItemsOptions(unittest.TestCase):
         # context_url contains something, otherwise this test is meaningless
         self.assertTrue(bool(context_url))
 
-        self.assertEquals(
+        self.assertEqual(
             options['rootUrl'],
             root_url
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['rootPath'],
             root_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['vocabularyUrl'],
             root_url + '/@@vocab?name=test_vocab&field=testfield'
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['basePath'],
             root_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['contextPath'],
             context_path
         )
 
-        self.assertEquals(
+        self.assertEqual(
             options['separator'],
             '#!@'
         )
@@ -274,3 +275,27 @@ class TestRelatedItemsOptions(unittest.TestCase):
         self.assertTrue(
             'favorites' not in options
         )
+
+
+class TestTinyMCEOptions(unittest.TestCase):
+    layer = PLONEAPPWIDGETS_INTEGRATION_TESTING
+
+    def setUp(self):
+        setRoles(self.layer['portal'], TEST_USER_ID, ['Contributor'])
+
+    def test__tinymce_options_different_contexts(self):
+        """Test if ``get_tinymce_options`` can be called with different
+        contexts, including invalid and form contexts.
+        """
+        request = self.layer['request']
+        portal = self.layer['portal']
+        portal.invokeFactory('Folder', 'sub')
+        sub = portal.sub
+
+        # TinyMCE on portal context
+        options = get_tinymce_options(portal, None, request)
+        self.assertEqual(options['relatedItems']['basePath'], '/plone')
+
+        # TinyMCE on sub folder context
+        options = get_tinymce_options(sub, None, request)
+        self.assertEqual(options['relatedItems']['basePath'], '/plone/sub')