+++ CHANGES.rst
@@ -20,7 +20,8 @@ New features:
 
 Bug fixes:
 
-- *add item here*
+- Change deprecated unittest method ``assertEquals`` to ``assertEqual``.
+  [thet]
 
 
 2.1 (2017-02-20)